Refactor F1 command to class-based structure
Replaces the old function-based F1 command with a new class-based implementation in Commands/F1.php, improving code organization and maintainability. Removes the old F1 command from Commands/OldCommands and updates CommandHandler.php to reflect the change.
